src/features/Home/: Clean up Home
